refactor: move ArrayOf to utils
